server/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
psycopg2-binary
python-dotenv
Werkzeug
PyYAML
pytest
